Add MetadataViewImplProxy as a built-in (for net45)
