+++ src/tests/Microsoft.VisualStudio.Composition.Tests/AssembliesLazyLoadedTests.cs
@@ -152,8 +152,7 @@ namespace Microsoft.VisualStudio.Composition.Tests
         public async Task ComposableAssembliesLazyLoadedByLazyTMetadata()
         {
             var catalog = TestUtilities.EmptyCatalog.AddParts(
-                await TestUtilities.V2Discovery.CreatePartsAsync(typeof(PartThatLazyImportsExportWithTypeMetadataViaTMetadata), typeof(AnExportWithMetadataTypeValue)))
-                .WithDesktopSupport();
+                await TestUtilities.V2Discovery.CreatePartsAsync(typeof(PartThatLazyImportsExportWithTypeMetadataViaTMetadata), typeof(AnExportWithMetadataTypeValue)));
             var catalogCache = await this.SaveCatalogAsync(catalog);
             var configuration = CompositionConfiguration.Create(catalog);
             var compositionCache = await this.SaveConfigurationAsync(configuration);
@@ -180,8 +179,7 @@ namespace Microsoft.VisualStudio.Composition.Tests
         public async Task ComposableAssembliesLazyLoadedWithGenericTypeArg()
         {
             var catalog = TestUtilities.EmptyCatalog.AddParts(
-                await TestUtilities.V2Discovery.CreatePartsAsync(typeof(PartImportingOpenGenericExport), typeof(OpenGenericExport<>)))
-                .WithDesktopSupport();
+                await TestUtilities.V2Discovery.CreatePartsAsync(typeof(PartImportingOpenGenericExport), typeof(OpenGenericExport<>)));
             var catalogCache = await this.SaveCatalogAsync(catalog);
             var configuration = CompositionConfiguration.Create(catalog);
             var compositionCache = await this.SaveConfigurationAsync(configuration);
